requirements.txt
PyYAML == 6.0
uritemplate == 4.1.1
Markdown == 3.4.1
Pygments == 2.12.0
django-filter == 22.1
django-guardian == 2.4.0
djangorestframework == 3.13.1
django-extensions==3.2.0
Django ==4.0